POST /users/ and DELETE /users/{username}/
This one got away from me. It started off as cleaning up Swagger documentation, then spiralled off into new endpoints, permission classes, a custom exception handler. It's still a bit of a tangled mess, but it's time to commit this and get back to the good, atomic commits that my mama raised me to make.